Split WAL inspection into dedicated CLI
